chore: moved shared generator type into shared type file
